conda_requirements.txt
wheel
Pillow
numpy
umsgpack
Cython
tqdm
scikit-image
scikit-learn 
pyyaml
flake8
setuptools
tensorboardX
einops
timm
